refactor ibd logic to modules
