+++ protocol/flows/src/v5/ibd/streams.rs
@@ -1,0 +1,116 @@
+//!
+//! Logical stream abstractions used throughout the IBD negotiation protocols
+//!
+
+use consensus_core::header::Header;
+use kaspa_core::{debug, info};
+use p2p_lib::{
+    common::FlowError,
+    convert::model::trusted::TrustedDataEntry,
+    make_message,
+    pb::{kaspad_message::Payload, RequestNextHeadersMessage, RequestNextPruningPointAndItsAnticoneBlocksMessage},
+    IncomingRoute, Router,
+};
+use std::sync::Arc;
+
+const IBD_BATCH_SIZE: usize = 99;
+
+pub struct TrustedEntryStream<'a, 'b> {
+    router: &'a Router,
+    incoming_route: &'b mut IncomingRoute,
+    i: usize,
+}
+
+impl<'a, 'b> TrustedEntryStream<'a, 'b> {
+    pub fn new(router: &'a Router, incoming_route: &'b mut IncomingRoute) -> Self {
+        Self { router, incoming_route, i: 0 }
+    }
+
+    pub async fn next(&mut self) -> Result<Option<TrustedDataEntry>, FlowError> {
+        let msg = match tokio::time::timeout(p2p_lib::common::DEFAULT_TIMEOUT, self.incoming_route.recv()).await {
+            Ok(op) => {
+                if let Some(msg) = op {
+                    match msg.payload {
+                        Some(Payload::BlockWithTrustedDataV4(payload)) => Ok(Some(payload.try_into()?)),
+                        Some(Payload::DoneBlocksWithTrustedData(_)) => {
+                            debug!("trusted entry stream completed after {} items", self.i);
+                            return Ok(None);
+                        }
+                        _ => Err(FlowError::UnexpectedMessageType(
+                            stringify!(Payload::BlockWithTrustedDataV4 | Payload::DoneBlocksWithTrustedData),
+                            Box::new(msg.payload),
+                        )),
+                    }
+                } else {
+                    Err(FlowError::P2pConnectionError(p2p_lib::ConnectionError::ChannelClosed))
+                }
+            }
+            Err(_) => Err(FlowError::Timeout(p2p_lib::common::DEFAULT_TIMEOUT)),
+        };
+
+        // Request the next batch
+        // TODO: test that batch counting is correct and follows golang imp
+        self.i += 1;
+        if self.i % IBD_BATCH_SIZE == 0 {
+            info!("Downloaded {} blocks from the pruning point anticone", self.i - 1);
+            self.router
+                .enqueue(make_message!(
+                    Payload::RequestNextPruningPointAndItsAnticoneBlocks,
+                    RequestNextPruningPointAndItsAnticoneBlocksMessage {}
+                ))
+                .await?;
+        }
+
+        msg
+    }
+}
+
+/// A chunk of headers
+pub type HeadersChunk = Vec<Arc<Header>>;
+
+pub struct HeadersChunkStream<'a, 'b> {
+    router: &'a Router,
+    incoming_route: &'b mut IncomingRoute,
+    i: usize,
+}
+
+impl<'a, 'b> HeadersChunkStream<'a, 'b> {
+    pub fn new(router: &'a Router, incoming_route: &'b mut IncomingRoute) -> Self {
+        Self { router, incoming_route, i: 0 }
+    }
+
+    pub async fn next(&mut self) -> Result<Option<HeadersChunk>, FlowError> {
+        let msg = match tokio::time::timeout(p2p_lib::common::DEFAULT_TIMEOUT, self.incoming_route.recv()).await {
+            Ok(op) => {
+                if let Some(msg) = op {
+                    match msg.payload {
+                        Some(Payload::BlockHeaders(payload)) => {
+                            if payload.block_headers.is_empty() {
+                                // The syncer should have sent a done message if the search completed, and not an empty list
+                                return Err(FlowError::ProtocolError("Received an empty headers message"));
+                            }
+                            Ok(Some(payload.try_into()?))
+                        }
+                        Some(Payload::DoneHeaders(_)) => {
+                            debug!("headers chunk stream completed after {} chunks", self.i);
+                            return Ok(None);
+                        }
+                        _ => Err(FlowError::UnexpectedMessageType(
+                            stringify!(Payload::BlockHeaders | Payload::DoneHeaders),
+                            Box::new(msg.payload),
+                        )),
+                    }
+                } else {
+                    Err(FlowError::P2pConnectionError(p2p_lib::ConnectionError::ChannelClosed))
+                }
+            }
+            Err(_) => Err(FlowError::Timeout(p2p_lib::common::DEFAULT_TIMEOUT)),
+        };
+
+        // Request the next chunk
+        self.i += 1;
+        self.router.enqueue(make_message!(Payload::RequestNextHeaders, RequestNextHeadersMessage {})).await?;
+
+        msg
+    }
+}
