+++ protocol/p2p/src/convert/mod.rs
@@ -4,5 +4,6 @@ pub mod ghostdag;
 pub mod hash;
 pub mod header;
 pub mod messages;
+pub mod model;
 pub mod option;
 pub mod pruning;
